+++ CHANGELOG.md
@@ -13,11 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ESP32-C6: Properly initialize PMU (#974)
 - Implement overriding base mac address (#1044)
+- Add `rt-riscv` and `rt-xtensa` features to enable/disable runtime support (#1057)
 - ESP32-C6: Implement deep sleep (#918)
 
 ### Changed
 
 ### Fixed
+
+- ESP32: correct gpio 32/33 in errata36() (#1053)
 
 ### Removed